[flink] Preserve large pushed-down limits
Forward Flink long limits without narrowing and cover values above the signed integer range

+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/source/FlinkSourceBuilder.java
@@ -212,7 +212,7 @@ public class FlinkSourceBuilder {
             readBuilder.withPartitionFilter(partitionPredicate);
         }
         if (limit != null) {
-            readBuilder.withLimit(limit.intValue());
+            readBuilder.withLimit(limit);
         }
         return readBuilder.dropStats();
     }
